Standardise config and support use of servicebus queues in azure
