+++ src/SFA.DAS.EmployerApprenticeshipsService.Web/packages.config
@@ -49,11 +49,13 @@
   <package id="SFA.DAS.Configuration.AzureTableStorage" version="1.0.0.4757" targetFramework="net45" />
   <package id="SFA.DAS.Configuration.FileStorage" version="1.0.0.4757" targetFramework="net45" />
   <package id="SFA.DAS.Messaging" version="2.0.0.6029" targetFramework="net45" />
+  <package id="SFA.DAS.Messaging.AzureServiceBus" version="2.0.0.5827" targetFramework="net45" />
   <package id="structuremap" version="3.1.1.134" targetFramework="net45" />
   <package id="StructureMap.MVC5" version="3.1.1.134" targetFramework="net45" />
   <package id="structuremap.web" version="3.1.0.133" targetFramework="net45" />
   <package id="System.Spatial" version="5.7.0" targetFramework="net45" />
   <package id="WebActivatorEx" version="2.0.5" targetFramework="net45" />
   <package id="WebGrease" version="1.5.2" targetFramework="net45" />
+  <package id="WindowsAzure.ServiceBus" version="3.1.7" targetFramework="net45" />
   <package id="WindowsAzure.Storage" version="6.2.0" targetFramework="net45" />
 </packages>
