--- src/SFA.DAS.LevyAggregationProvider.Worker/LevyAggregationConfiguration.cs
@@ -1,12 +1,0 @@
-﻿namespace SFA.DAS.LevyAggregationProvider.Worker
-{
-    public class LevyAggregationConfiguration
-    {
-        public EmployerConfiguration Employer { get; set; }
-    }
-
-    public class EmployerConfiguration
-    {
-        public string DatabaseConnectionString { get; set; }
-    }
-}
