+++ src/SFA.DAS.LevyAggregationProvider.Worker/packages.config
@@ -11,7 +11,9 @@
   <package id="SFA.DAS.Configuration.AzureTableStorage" version="1.0.0.4757" targetFramework="net45" />
   <package id="SFA.DAS.Configuration.FileStorage" version="1.0.0.4757" targetFramework="net45" />
   <package id="SFA.DAS.Messaging" version="2.0.0.5827" targetFramework="net45" />
+  <package id="SFA.DAS.Messaging.AzureServiceBus" version="2.0.0.5827" targetFramework="net45" />
   <package id="structuremap" version="4.2.0.402" targetFramework="net45" />
   <package id="System.Spatial" version="5.7.0" targetFramework="net45" />
+  <package id="WindowsAzure.ServiceBus" version="3.1.7" targetFramework="net45" />
   <package id="WindowsAzure.Storage" version="6.2.0" targetFramework="net45" />
 </packages>
