--- src/SFA.DAS.LevyDeclarationProvider.Worker/EmployerApprenticeshipsServiceConfiguration.cs
@@ -1,18 +1,0 @@
-﻿namespace SFA.DAS.LevyDeclarationProvider.Worker
-{
-    public class EmployerApprenticeshipsServiceConfiguration
-    {
-        public CompaniesHouseConfiguration CompaniesHouse { get; set; }
-        public EmployerConfiguration Employer { get; set; }
-    }
-
-    public class EmployerConfiguration
-    {
-        public string DatabaseConnectionString { get; set; }
-    }
-
-    public class CompaniesHouseConfiguration
-    {
-        public string ApiKey { get; set; }
-    }
-}
