+++ src/SFA.DAS.LevyDeclarationProvider.Worker/SFA.DAS.LevyDeclarationProvider.Worker.csproj
@@ -116,7 +116,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DependencyResolution\DefaultRegistry.cs" />
-    <Compile Include="EmployerApprenticeshipsServiceConfiguration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Providers\ILevyDeclaration.cs" />
     <Compile Include="Providers\LevyDeclaration.cs" />
@@ -140,7 +139,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SFA.DAS.EmployerApprenticeshipsService.Application\SFA.DAS.EmployerApprenticeshipsService.Application.csproj">
-     <Project>{0226F97A-704E-49B0-A100-B75FEEC9C8AD}</Project>
+      <Project>{0226F97A-704E-49B0-A100-B75FEEC9C8AD}</Project>
       <Name>SFA.DAS.EmployerApprenticeshipsService.Application</Name>
     </ProjectReference>
     <ProjectReference Include="..\SFA.DAS.EmployerApprenticeshipsService.Domain\SFA.DAS.EmployerApprenticeshipsService.Domain.csproj">
